change idx types from int32 to int64 (#101)
Conflicts:
	python/mxnet/test_utils.py
	tests/python/unittest/test_sparse_operator.py

update mshadow submodule

fix extra quotes in test script

change indptr type to int64

better err message for rsp"

+++ python/mxnet/test_utils.py
@@ -76,7 +76,6 @@ def random_sample(population, k):
     return population_copy[0:k]
 
 
-# TODO(haibin) also include types in arguments
 def rand_sparse_ndarray(shape, storage_type, density=None):
     """Generate a random sparse ndarray. Returns the ndarray, value(np) and indices(np) """
     density = rnd.rand() if density is None else density
@@ -90,10 +89,10 @@ def rand_sparse_ndarray(shape, storage_type, density=None):
         indices = np.argwhere(idx_sample < density).flatten()
         if indices.shape[0] == 0:
             result = mx.nd.zeros(shape, storage_type='row_sparse')
-            return result, (np.array([]), np.array([], dtype='int32'))
+            return result, (np.array([], dtype='int64'), np.array([], dtype='int64'))
         # generate random values
         val = rnd.rand(indices.shape[0], num_cols)
-        arr = mx.sparse_nd.row_sparse(val, indices, shape, indices_type=np.int32)
+        arr = mx.sparse_nd.row_sparse(val, indices, shape, indices_type=np.int64)
         return arr, (val, indices)
     elif storage_type == 'csr':
         assert(len(shape) == 2)
